backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
gspread==6.1.2
google-auth==2.34.0
pydantic==2.8.2
pandas==2.2.2
scikit-learn==1.5.2
python-dotenv==1.0.1
